requirements.txt
asgiref==3.6.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
cloudinary==1.30.0
dj-database-url==1.2.0
dj-rest-auth==2.2.5
Django==3.2.16
django-allauth==0.50.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
mypy-extensions==0.4.3
oauthlib==3.2.2
pathspec==0.10.3
Pillow==9.3.0
psycopg2==2.9.5
pycodestyle==2.10.0
pydot==1.4.2
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.7.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
